backend/requirements.txt
# Core LLM & Agent Tools
openai==1.84.0
langchain==0.3.25
together==1.2.3
tiktoken==0.9.0
groq==0.26.0

# API & Web Integration
requests==2.32.3
aiohttp==3.12.9
python-dotenv==1.1.0

# Multilingual NLP
langdetect==1.0.9
deep-translator==1.11.4

# Data Validation & Parsing
pydantic==2.7.4

# Logging & Testing
loguru==0.7.2
pytest==8.2.0

# Optional: Async HTTP or enhanced agent support
httpx==0.27.0

sentence-transformers
newsapi-python
langchain
faiss-cpu
openai
langchain_huggingface
pvporcupine
pyaudio
numpy
torch
whisper